ai_agent/requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.3.0
anthropic==0.7.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.1
numpy==1.24.3
python-dotenv==1.0.0
pydantic==2.4.2
tiktoken==0.5.1
aiohttp==3.8.6
websockets==11.0.3
json5==0.9.14
markdown==3.5.1
rich==13.5.2
click==8.1.7
schedule==1.2.0
watchdog==3.0.0
pytz==2023.3
python-dateutil==2.8.2